# Conflicts:
#	src/MolSim.cpp
#	src/ParticleGenerator.cpp
#	src/ParticleGenerator.h
#	src/Simulation/MoleculeSimulation.cpp
#	src/Simulation/MoleculeSimulation.h
#	src/Simulation/PlanetSimulation.cpp
#	src/Simulation/PlanetSimulation.h
#	src/Simulation/SimulationFactory.cpp
#	src/Simulation/SimulationFactory.h
#	src/inputReader/Arguments.h
#	src/inputReader/FileReader.cpp
#	src/inputReader/FileReader.h
#	src/inputReader/FileReaderCuboids.h
#	src/inputReader/InputReader.cpp
#	tests/behaviour/ParticleGenerator.cpp

+++ src/inputReader/Cuboid.h
@@ -3,7 +3,7 @@
 // Struct with the given parameters of the cuboid and a container for storing the particles
 struct Cuboid {
   std::array<double, 3> origin{};
-  std::array<size_t, 3> numPerDim{};
+  std::array<int, 3> numPerDim{};
   double h{0.0};
   double mass{1.0};
   std::array<double, 3> baseVelocity{};
